+++ cpp/src/io/soa/polygon_soa_reader.cu
@@ -23,7 +23,7 @@
 #include <rmm/rmm.h>
 #include <utilities/error_utils.hpp>
 #include <cuspatial/soa_readers.hpp>
-#include <cuspatial/shared_util.h>
+#include <utility/utility.hpp>
 
 namespace cuspatial
 {
