+++ cpp/src/spatial/hausdorff.cu
@@ -21,7 +21,7 @@
 #include <sys/time.h>
 #include <time.h>
 
-#include <cuspatial/shared_util.h>
+#include <utility/utility.hpp>
 #include <cuspatial/hausdorff.hpp>
 
 using namespace std;
