+++ cpp/src/utility/st_query_thrust.cuh
