+++ cpp/src/utility/trajectory_thrust.cuh
