+++ cpp/src/utility/utility.hpp
