+++ cpp/tests/spatial/hausdorff_compare.cu
@@ -25,7 +25,7 @@
 
 #include <cuspatial/soa_readers.hpp> 
 #include <cuspatial/hausdorff.hpp> 
-#include <cuspatial/shared_util.h> 
+#include <utility/utility.hpp> 
 #include "hausdorff_util.h" 
 
 #include <tests/utilities/cudf_test_utils.cuh>
