+++ cpp/tests/spatial/hausdorff_test.cu
@@ -22,7 +22,7 @@
 
 #include <cuspatial/soa_readers.hpp> 
 #include <cuspatial/hausdorff.hpp> 
-#include <cuspatial/shared_util.h> 
+#include <utility/utility.hpp> 
 #include "hausdorff_util.h" 
 
 #include <tests/utilities/cudf_test_utils.cuh>
